Define trait `ConcreteReader` for blanket `impl`s
The trait `ConcreteReader` defines a method `read_concrete` which
returns a reference to the concrete part of a concept. Blanket
implementations of `GetDefinitionOf` and `FindWhatReducesToIt` can
therefore be defined for types that implement `ConcreteReader`.
This allows different variants of `Concept` to be written which
expand on `ConcreteConcept` without having to reimplement
`GetDefinitionOf` or `FindWhatReducesToIt`.