Define trait `ConcreteWriter` for blanket `impl`s
The trait `ConcreteWriter` defines a method `write_concrete` which
returns a mutable reference to the concrete part of a concept. Blanket
implementations of `SetAsDefinitionOf`, `MakeReduceFrom`,
`RemoveAsDefinitionOf` and `NoLongerReducesFrom` can therefore be
defined for types that implement `ConcreteWriter`. This allows
different variants of `Concept` to be written which expand on
`ConcreteConcept` without having to reimplement the above traits.