Modify concept types so they actually make sense
Terminology was confusing due to way the `Concept` type was structured.

Change `Concept` from an `enum` to a `struct` with a `CommonPart` and
`SpecificPart` field.
- `CommonPart` is the same as the previous `ConcreteConcept` but is
a better name as it contains the common part of any type of concept.
- `SpecificPart` is an enum with the variants:
* `Concrete` for concepts that can't be defined as a composition of
concepts or reduce to other concepts.
* `Abstract(AbstractPart)` where `AbstractPart` is the previous
`AbstractConcept` without the `concrete_concept` field.
* `String(String)` for concrete concepts with an associated `String`
value.

Remove `StringConcept` due to redundancy.

Remove the traits `ConcreteReader` and `ConcreteWriter` due to the
above changes making these redundant.