Change tests that set reductions and definitions
Need a new primitive that allows the distinction between the querying
the truth of `a (-> b)` or `a (:= (b c))` and setting these as facts.

The tests have been changed so that `let (a (-> b))` should set a
reduction rule and `let (a (:= (b c)))` should set a definition.

Most of the tests obviously fail now.